Cool Boost stuff
Player now carries momentum when stopping, even if their speed is higher than the maxSpeed and they are moving, also boosting is a thing now. You also have a lower deceleration rate in the air vs the ground.
Tiles now only update when they are int a position that they can be checked in, although this can cause tearing in the stage, but it could fit the cyber aesthetic so idk