Optimized at last
Reduce the area of collision to improve performance. Also added double jump but I think I accidentally removed it, whoops, but idk if I even like it so I might not even re add it.
Level tiles are also trimmed now, so only the ones on screen are taken into account when checking all tiles. Added a new tile that sends the player to the left, and updated tile 6 to do the same but to the right , instead of it's old behaviour of sending the player in their current direction.